+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/common/components/RemindAtRow.kt
@@ -38,7 +38,7 @@ fun ColumnScope.RemindAtRow(
 ) {
     Row(
         horizontalArrangement = Arrangement.SpaceBetween,
-        modifier = Modifier
+        modifier = modifier
             .fillMaxWidth()
             .background(MaterialTheme.colors.surface)
     ) {
