+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/common/components/TimeDateRow.kt
@@ -1,4 +1,4 @@
-package com.realityexpander.tasky.agenda_feature.presentation.components
+package com.realityexpander.tasky.agenda_feature.presentation.common.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
