+++ app/src/main/java/com/realityexpander/tasky/agenda_feature/presentation/event_screen/EventScreen.kt
@@ -32,9 +32,9 @@ import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
 import com.realityexpander.tasky.agenda_feature.domain.AgendaItem
 import com.realityexpander.tasky.agenda_feature.domain.Attendee
+import com.realityexpander.tasky.agenda_feature.presentation.common.components.EditTextModal
+import com.realityexpander.tasky.agenda_feature.presentation.common.components.TimeDateRow
 import com.realityexpander.tasky.agenda_feature.presentation.common.util.toZonedDateTime
-import com.realityexpander.tasky.agenda_feature.presentation.components.EditTextModal
-import com.realityexpander.tasky.agenda_feature.presentation.components.TimeDateRow
 import com.realityexpander.tasky.agenda_feature.presentation.event_screen.EventScreenEvent.*
 import com.realityexpander.tasky.agenda_feature.presentation.event_screen.components.AttendeeList
 import com.realityexpander.tasky.agenda_feature.presentation.event_screen.components.PillButton
@@ -172,7 +172,7 @@ fun AddEventScreenContent(
                     .weight(1f)
             )
 
-            // • Edit / Save Button
+            // • EDIT / SAVE BUTTON
             if (isEditable) {
                 Text(
                     text = "Save",
@@ -510,142 +510,12 @@ fun AddEventScreenContent(
                         )
                     }
                 )
-
-//                Row(
-//                    horizontalArrangement = Arrangement.SpaceBetween,
-//                    modifier = Modifier
-//                        .fillMaxWidth()
-//                        .background(MaterialTheme.colors.surface)
-//                ) {
-//
-//                    Row(
-//                        horizontalArrangement = Arrangement.SpaceBetween,
-//                        modifier = Modifier
-//                            .weight(1f)
-//                            .clickable(enabled = isEditable) {
-//                                onAction(
-//                                    SetEditMode(
-//                                        EditMode.RemindAtDateTime(
-//                                            state.event?.remindAt ?: ZonedDateTime.now()
-//                                        )
-//                                    )
-//                                )
-//                            }
-//                    ) {
-//                        Row(
-//                            horizontalArrangement = Arrangement.Start,
-//                            modifier = Modifier
-//                                .weight(1f)
-//                                .padding(start = DP.small, end = DP.medium)
-//                        ) {
-//                            // • ALARM/REMINDER AT ICON
-//                            Icon(
-//                                imageVector = Icons.Outlined.Notifications,
-//                                tint = MaterialTheme.colors.onSurface.copy(alpha = .3f),
-//                                contentDescription = "Remind At Button",
-//                                modifier = Modifier
-//                                    .size(34.dp)
-//                                    .clip(shape = RoundedCornerShape(5.dp))
-//                                    .background(MaterialTheme.colors.onSurface.copy(alpha = .1f))
-//                                    .padding(4.dp)
-//                                    .align(Alignment.CenterVertically)
-//                                    .weight(.2f)
-//                            )
-//                            Spacer(modifier = Modifier.smallWidth())
-//                            Text(
-//                                state.event?.from?.toTimeDifferenceHumanReadable(state.event.remindAt)
-//                                    ?: "not set",
-//                                color = MaterialTheme.colors.onSurface,
-//                                textAlign = TextAlign.Start,
-//                                modifier = Modifier
-//                                    .align(Alignment.CenterVertically)
-//                                    .weight(1f)
-//                            )
-//                        }
-//
-//                        Icon(
-//                            imageVector = Icons.Filled.ChevronRight,
-//                            tint = if (isEditable) MaterialTheme.colors.onSurface else Color.Transparent,
-//                            contentDescription = "Edit Remind At DateTime",
-//                            modifier = Modifier
-//                                .weight(.1125f)
-//                                .align(Alignment.CenterVertically)
-//                        )
-//
-//                        // • REMIND AT MENU
-//                        DropdownMenu(
-//                            expanded = state.editMode is EditMode.RemindAtDateTime,
-//                            offset = DpOffset(DP.small, 0.dp),
-//                            onDismissRequest = { onAction(CancelEditMode) },
-//                            modifier = Modifier
-//                                .background(color = MaterialTheme.colors.onSurface)
-//
-//                        ) {
-//                            MenuItem(
-//                                title = "10 minutes before",
-//                                onClick = {
-//                                    onAction(
-//                                        EditMode.SaveDateTime(
-//                                            state.event?.from?.minusMinutes(10)
-//                                            ?: ZonedDateTime.now()
-//                                        )
-//                                    )
-//                                },
-//                            )
-//                            MenuItem(
-//                                title = "30 minutes before",
-//                                onClick = {
-//                                    onAction(
-//                                        EditMode.SaveDateTime(
-//                                        state.event?.from?.minusMinutes(30)
-//                                            ?: ZonedDateTime.now()
-//                                        )
-//                                    )
-//                                },
-//                            )
-//                            MenuItem(
-//                                title = "1 hour before",
-//                                onClick = {
-//                                    onAction(
-//                                        EditMode.SaveDateTime(
-//                                        state.event?.from?.minusHours(1)
-//                                            ?: ZonedDateTime.now()
-//                                        )
-//                                    )
-//                                },
-//                            )
-//                            MenuItem(
-//                                title = "6 hours before",
-//                                onClick = {
-//                                    onAction(
-//                                        EditMode.SaveDateTime(
-//                                        state.event?.from?.minusHours(6)
-//                                            ?: ZonedDateTime.now()
-//                                        )
-//                                    )
-//                                },
-//                            )
-//                            MenuItem(
-//                                title = "1 day before",
-//                                onClick = {
-//                                    onAction(
-//                                        EditMode.SaveDateTime(
-//                                        state.event?.from?.minusDays(1)
-//                                            ?: ZonedDateTime.now()
-//                                        )
-//                                    )
-//                                },
-//                            )
-//                        }
-//            }
-
-//        }
                 SmallHeightHorizontalDivider()
                 Spacer(modifier = Modifier.smallHeight())
 
 
                 // • ATTENDEES HEADER (VISITORS & ADD ATTENDEE BUTTON)
-                Row() {
+                Row {
                     Text(
                         "Visitors",
                         color = MaterialTheme.colors.onSurface,
@@ -713,7 +583,6 @@ fun AddEventScreenContent(
                             attendeeListType = AttendeeListType.NOT_GOING
                         }
                     )
-//                    Spacer(modifier = Modifier.smallWidth())
                 }
                 Spacer(modifier = Modifier.mediumHeight())
 
@@ -754,8 +623,8 @@ fun AddEventScreenContent(
                 // • JOIN/LEAVE/DELETE EVENT BUTTON
                 Text(
                     if (state.event?.isUserEventCreator == true) "DELETE EVENT"
-                    else if (state.event?.isGoing == true) "LEAVE EVENT"
-                    else "JOIN EVENT",
+                        else if (state.event?.isGoing == true) "LEAVE EVENT"
+                        else "JOIN EVENT",
                     style = MaterialTheme.typography.h4,
                     color = MaterialTheme.colors.onSurface.copy(alpha = 0.3f),
                     textAlign = TextAlign.Center,
@@ -766,92 +635,91 @@ fun AddEventScreenContent(
 
             }
 
-                // • EDITORS FOR EVENT PROPERTIES
-                state.editMode?.let { editMode ->
-                    when (editMode) {
-                        is EditMode.TitleText,
-                        is EditMode.DescriptionText -> {
-                            editMode as EditMode.EditText
-                            EditTextModal(
-                                text = editMode.text,
-                                title = editMode.title,
-                                editTextStyle =
-                                if (editMode is EditMode.TitleText)
-                                    MaterialTheme.typography.h2  // cant access from non-compose function, make a wrapper?
-                                else
-                                    MaterialTheme.typography.body1, // cant access from non-compose function, make a wrapper?
-                                onSave = {
-                                    onAction(EditMode.SaveText(it))
-                                },
-                                onCancel = {
+            // • EDITORS FOR EVENT PROPERTIES
+            state.editMode?.let { editMode ->
+                when (editMode) {
+                    is EditMode.TitleText,
+                    is EditMode.DescriptionText -> {
+                        editMode as EditMode.EditText
+                        EditTextModal(
+                            text = editMode.text,
+                            title = editMode.title,
+                            editTextStyle =
+                            if (editMode is EditMode.TitleText)
+                                MaterialTheme.typography.h2  // cant access from non-compose function, make a wrapper?
+                            else
+                                MaterialTheme.typography.body1, // cant access from non-compose function, make a wrapper?
+                            onSave = {
+                                onAction(EditMode.SaveText(it))
+                            },
+                            onCancel = {
+                                onAction(CancelEditMode)
+                            }
+                        )
+                    }
+                    is EditMode.FromDate,
+                    is EditMode.ToDate -> {
+                        editMode as EditMode.EditDateTime
+                        var pickedDate by remember { mutableStateOf(LocalDateTime.now()) }
+                        val dateDialogState = rememberMaterialDialogState()
+                        dateDialogState.show()
+
+                        MaterialDialog(
+                            dialogState = dateDialogState,
+                            buttons = {
+                                positiveButton(text = "OK") {
+                                    onAction(EditMode.SaveDateTime(pickedDate?.toZonedDateTime()!!))
+                                }
+                                negativeButton(text = "CANCEL") {
+                                    dateDialogState.hide()
                                     onAction(CancelEditMode)
                                 }
-                            )
-                        }
-                        is EditMode.FromDate,
-                        is EditMode.ToDate -> {
-                            editMode as EditMode.EditDateTime
-                            var pickedDate by remember { mutableStateOf(LocalDateTime.now()) }
-                            val dateDialogState = rememberMaterialDialogState()
-                            dateDialogState.show()
-
-                            MaterialDialog(
-                                dialogState = dateDialogState,
-                                buttons = {
-                                    positiveButton(text = "OK") {
-                                        onAction(EditMode.SaveDateTime(pickedDate?.toZonedDateTime()!!))
-                                    }
-                                    negativeButton(text = "CANCEL") {
-                                        dateDialogState.hide()
-                                        onAction(CancelEditMode)
-                                    }
-                                }
+                            }
+                        ) {
+                            datepicker(
+                                initialDate = editMode.dateTime.toLocalDate(),
+                                title = editMode.title,
                             ) {
-                                datepicker(
-                                    initialDate = editMode.dateTime.toLocalDate(),
-                                    title = editMode.title,
-                                ) {
-                                    pickedDate = it.atTime(editMode.dateTime.toLocalTime())
-                                }
+                                pickedDate = it.atTime(editMode.dateTime.toLocalTime())
                             }
                         }
-                        is EditMode.FromTime,
-                        is EditMode.ToTime -> {
-                            editMode as EditMode.EditDateTime
-                            var pickedTime by remember { mutableStateOf(LocalDateTime.now()) }
-                            val dateDialogState = rememberMaterialDialogState()
-                            dateDialogState.show()
-
-                            MaterialDialog(
-                                dialogState = dateDialogState,
-                                buttons = {
-                                    positiveButton(text = "OK") {
-                                        onAction(EditMode.SaveDateTime(pickedTime?.toZonedDateTime()!!))
-                                    }
-                                    negativeButton(text = "CANCEL") {
-                                        dateDialogState.hide()
-                                        onAction(CancelEditMode)
-                                    }
-                                }
-                            ) {
-                                timepicker(
-                                    initialTime = editMode.dateTime.toLocalTime(),
-                                    title = editMode.title,
-                                ) {
-                                    pickedTime = it.atDate(editMode.dateTime.toLocalDate())
-                                }
-                            }
-                        }
-                        is EditMode.RemindAtDateTime -> { /* handled in the RemindAt UI element, this is here to remove compiler warning */
-                        }
-                        is EditMode.AddPhoto -> TODO()
-                        is EditMode.ConfirmDeletePhoto -> TODO()
-                        is EditMode.AddAttendee -> TODO()
-                        is EditMode.ConfirmDeleteAttendee -> TODO()
                     }
+                    is EditMode.FromTime,
+                    is EditMode.ToTime -> {
+                        editMode as EditMode.EditDateTime
+                        var pickedTime by remember { mutableStateOf(LocalDateTime.now()) }
+                        val dateDialogState = rememberMaterialDialogState()
+                        dateDialogState.show()
+
+                        MaterialDialog(
+                            dialogState = dateDialogState,
+                            buttons = {
+                                positiveButton(text = "OK") {
+                                    onAction(EditMode.SaveDateTime(pickedTime?.toZonedDateTime()!!))
+                                }
+                                negativeButton(text = "CANCEL") {
+                                    dateDialogState.hide()
+                                    onAction(CancelEditMode)
+                                }
+                            }
+                        ) {
+                            timepicker(
+                                initialTime = editMode.dateTime.toLocalTime(),
+                                title = editMode.title,
+                            ) {
+                                pickedTime = it.atDate(editMode.dateTime.toLocalDate())
+                            }
+                        }
+                    }
+                    is EditMode.RemindAtDateTime -> { /* handled in the RemindAt UI element, this is here to remove compiler warning */
+                    }
+                    is EditMode.AddPhoto -> TODO()
+                    is EditMode.ConfirmDeletePhoto -> TODO()
+                    is EditMode.AddAttendee -> TODO()
+                    is EditMode.ConfirmDeleteAttendee -> TODO()
                 }
             }
-//        }
+        }
     }
 }
